use-cases/04-agent-tool-use/python/pyproject.toml
[project]
name = "agent-tool-use"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=1.0.10",
    "langchain-openai>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "python-dotenv>=1.0.0",
]
